Fix TypeError on updating job configuration
This commit fixes a TypeError ("ItemView object is not a mapping")
that occurs when updating the Prometheus job configurations. As
part of the fix the relevant code has also be simplified and made
more explicit instead of depending on the fact the same key has
been declared multiple times in the construction of a dictionary
the last value of the key is the one that actually takes effect.