Refactor to remove multiplicity of simlar methods
The commit is the first step in a refactor to remove multiple
similar methods and replace then with methods that have clearly
distinct responsibilities. Prior to this commit there
were three methods that handled updating metrics consumers

1) _update_new_metrics_consumer()
2) _update_all_metrics_consumers()
3) _update_metrics_consumer_relations()

(2) and (3) updated relation data for all metrics consumers. (3)
accepted a list of metrics consumer relations and updated relation
data for all of them. (2) only set a status message and passed
a list of relations to (3). (1) passed a list of one relation to
(3).

The plan for this refactor is replace the three methods with two

A) _update_metrics_consumer_relation() that accepts a SINGLE relation
object and updates its relation data. This method replaces (1) and
(3)

B) _update_all_metrics_consumers() this iterates over all consumer
relations and updates them by invoking (A) on each of those
relations.