Remove alert `group` duplication  (#6)
* Renamed provider and consumers

This commit starts a refactor of this scrape config charm towards
fixing existing bugs. As part of this refactor this commit
renames variables and status messages to make them consistent
with the "Metrics Provider" and "Metrics Consumer" terminology
used by the Prometheus Charm.

* Renamed provider consumer relation predicates

This commit renames the provider/consumer relation checking
predictes to use shorter names. Also an unused property method
is being deleted.

* Configuration keys use underscores now

Previously the configuration keys used hyphens because juju
seemed to produce errors when underscores were used. However
the Prometheus configuration file used underscores. Hence
hyphens needed to be converted to underscores before fowarding
these configurations to Prometheus. Now that Juju accepts
underscores this commit removes all this logic prefering
to use underscores consistently as this is accepted policy
with the observability team and because Prometheus config
keys themselves use underscores.

* Fix TypeError on updating job configuration

This commit fixes a TypeError ("ItemView object is not a mapping")
that occurs when updating the Prometheus job configurations. As
part of the fix the relevant code has also be simplified and made
more explicit instead of depending on the fact the same key has
been declared multiple times in the construction of a dictionary
the last value of the key is the one that actually takes effect.

* Fix logic of relation existence predicates

This commit inverts the logic of the code in the predicate functions
that test for existence of relations, thereby making the correspondance
between the function name and encapsulated logic accurate.

* Removed duplicated nesting of alert rule groups

This commit removes duplicated nesting of alert rule groups. The
structure of "alert_rules" data is no
"{topology_identifier: {"groups": ... }}"

* Refactor to remove multiplicity of simlar methods

The commit is the first step in a refactor to remove multiple
similar methods and replace then with methods that have clearly
distinct responsibilities. Prior to this commit there
were three methods that handled updating metrics consumers

1) _update_new_metrics_consumer()
2) _update_all_metrics_consumers()
3) _update_metrics_consumer_relations()

(2) and (3) updated relation data for all metrics consumers. (3)
accepted a list of metrics consumer relations and updated relation
data for all of them. (2) only set a status message and passed
a list of relations to (3). (1) passed a list of one relation to
(3).

The plan for this refactor is replace the three methods with two

A) _update_metrics_consumer_relation() that accepts a SINGLE relation
object and updates its relation data. This method replaces (1) and
(3)

B) _update_all_metrics_consumers() this iterates over all consumer
relations and updates them by invoking (A) on each of those
relations.

* Strip topology key before forwarding alert rules

The alerts() method of MetricsEndpointConsumer returns
alert rules using the topology identifier as the key and
the actual alert rules relation data as the value. Hence
before the Prometheus Scrape Config charm needs to strip
this key before forwarding the alert rules relation data.

fix

* Revert obsolete callback introduced by rebase

This commit removes and obsolete and duplicated callback that
was introduced when rebasing onto main which had a recently
merged (but old) commit that was based on the state of the
main branch prior to starting work on this branch.

* Fix unit test

This commit fixes broken unit test due change in blocked status
messages.

* Fix linting in charm

* Updated docstrings and comments

* Added missing leader guard before updated metrics consumer

This commit adds a missing leader guard before updating metrics
consumer relation because this update requires setting application
data.

* Updated metrics consumers when provider departs

This commit ensures that all metrics consumers are informed
when a provider departs.

Co-authored-by: Balbir Thomas <balbir.thomas@canonical.com>